@@ -7,7 +7,6 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
     "dev:ssr": "ng run sample:serve-ssr",
     "serve:ssr": "node dist/sample/server/main.js",
     "build:ssr": "ng build --prod && ng run sample:server:production",
@@ -15,53 +14,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~9.1.0",
-    "@angular/common": "~9.1.0",
-    "@angular/compiler": "~9.1.0",
-    "@angular/core": "~9.1.0",
+    "@angular/animations": "~10.2.1",
+    "@angular/common": "~10.2.1",
+    "@angular/compiler": "~10.2.1",
+    "@angular/core": "~10.2.1",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~9.1.0",
-    "@angular/platform-browser": "~9.1.0",
-    "@angular/platform-browser-dynamic": "~9.1.0",
-    "@angular/platform-server": "~9.1.0",
-    "@angular/router": "~9.1.0",
-    "@angular/service-worker": "~9.1.0",
-    "@nguniversal/express-engine": "~9.1.0",
-    "firebase": "^7.13.2",
-    "proxy-polyfill": "^0.3.1",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "@angular/forms": "~10.2.1",
+    "@angular/platform-browser": "~10.2.1",
+    "@angular/platform-browser-dynamic": "~10.2.1",
+    "@angular/platform-server": "~10.2.1",
+    "@angular/router": "~10.2.1",
+    "@angular/service-worker": "^10.2.1",
+    "@nguniversal/express-engine": "~10.1.0",
+    "firebase": "^8.0.0",
+    "first-input-delay": "^0.1.3",
+    "proxy-polyfill": "^0.3.2",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.1",
+    "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "~0.901.0",
-    "@angular-devkit/build-angular": "~0.901.0",
-    "@angular/cli": "~9.1.0",
-    "@angular/compiler-cli": "~9.1.0",
-    "@angular/language-service": "~9.1.0",
-    "@nguniversal/builders": "^9.1.0",
+    "@angular-devkit/architect": "~0.1001.3",
+    "@angular-devkit/build-angular": "~0.1002.0",
+    "@angular/cli": "~10.2.0",
+    "@angular/compiler-cli": "~10.2.1",
+    "@angular/language-service": "~10.2.1",
+    "@firebase/app-types": "^0.6.1",
+    "@nguniversal/builders": "^10.1.0",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
+    "@types/node": "^14.11.2",
+    "codelyzer": "^6.0.0",
+    "express": "^4.17.1",
     "express-serve-static-core": "^0.1.1",
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.0",
-    "firebase-functions-test": "^0.2.0",
-    "firebase-tools": "^8.0.0",
+    "firebase-admin": "^8.13.0",
+    "firebase-functions": "^3.11.0",
+    "firebase-functions-test": "^0.2.2",
+    "firebase-tools": "^8.11.1",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
     "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.3",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "ts-node": "~9.0.0",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7"
   }
 }
