@@ -1,7 +1,7 @@
 import { Inject, Injectable, NgZone, Optional, PLATFORM_ID } from '@angular/core';
-import { messaging } from 'firebase/app';
-import { concat, EMPTY, Observable, of, throwError } from 'rxjs';
-import { catchError, defaultIfEmpty, map, mergeMap, observeOn, switchMap } from 'rxjs/operators';
+import firebase from 'firebase/app';
+import { concat, EMPTY, Observable, of, throwError, fromEvent } from 'rxjs';
+import { catchError, defaultIfEmpty, map, mergeMap, observeOn, switchMap, switchMapTo, shareReplay, filter, subscribeOn } from 'rxjs/operators';
 import {
   FIREBASE_APP_NAME,
   FIREBASE_OPTIONS,
@@ -14,7 +14,7 @@ import {
 } from '@angular/fire';
 import { isPlatformServer } from '@angular/common';
 
-export interface AngularFireMessaging extends Omit<ɵPromiseProxy<messaging.Messaging>, 'deleteToken' | 'getToken' | 'requestPermission'> {
+export interface AngularFireMessaging extends Omit<ɵPromiseProxy<firebase.messaging.Messaging>, 'deleteToken' | 'getToken' | 'requestPermission'> {
 }
 
 @Injectable({
@@ -39,50 +39,50 @@ export class AngularFireMessaging {
     const schedulers = new ɵAngularFireSchedulers(zone);
 
     const messaging = of(undefined).pipe(
-      observeOn(schedulers.outsideAngular),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
       switchMap(() => isPlatformServer(platformId) ? EMPTY : import('firebase/messaging')),
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
-      map(app => app.messaging())
+      map(app => app.messaging()),
+      shareReplay({ bufferSize: 1, refCount: false })
     );
 
-    if (!isPlatformServer(platformId)) {
-
-      this.requestPermission = messaging.pipe(
-        observeOn(schedulers.outsideAngular),
-        // tslint:disable-next-line
-        switchMap(messaging => messaging.requestPermission())
-      );
-
-    } else {
-
-      this.requestPermission = throwError('Not available on server platform.');
-
-    }
+    this.requestPermission = messaging.pipe(
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      // tslint:disable-next-line
+      switchMap(messaging => firebase.messaging.isSupported() ? messaging.requestPermission() : throwError('Not supported.'))
+    );
 
     this.getToken = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => messaging.getToken()),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() && Notification.permission === 'granted' ? messaging.getToken() : EMPTY),
       defaultIfEmpty(null)
     );
 
     const tokenChanges = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => new Observable(messaging.onTokenRefresh.bind(messaging)).pipe(
-        switchMap(() => messaging.getToken())
-      ))
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? new Observable<string>(emitter =>
+        messaging.onTokenRefresh(emitter.next, emitter.error, emitter.complete)
+      ) : EMPTY),
+      switchMapTo(this.getToken)
     );
 
-    this.tokenChanges = concat(
-      messaging.pipe(
-        observeOn(schedulers.outsideAngular),
-        switchMap(messaging => messaging.getToken())
-      ),
-      tokenChanges
+    this.tokenChanges = messaging.pipe(
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? concat(this.getToken, tokenChanges) : EMPTY)
     );
+
 
     this.messages = messaging.pipe(
-      observeOn(schedulers.outsideAngular),
-      switchMap(messaging => new Observable(messaging.onMessage.bind(messaging)))
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
+      switchMap(messaging => firebase.messaging.isSupported() ? new Observable<string>(emitter =>
+        messaging.onMessage(next => emitter.next(next), err => emitter.error(err), () => emitter.complete())
+      ) : EMPTY),
     );
 
     this.requestToken = of(undefined).pipe(
@@ -92,7 +92,8 @@ export class AngularFireMessaging {
     );
 
     this.deleteToken = (token: string) => messaging.pipe(
-      observeOn(schedulers.outsideAngular),
+      subscribeOn(schedulers.outsideAngular),
+      observeOn(schedulers.insideAngular),
       switchMap(messaging => messaging.deleteToken(token)),
       defaultIfEmpty(false)
     );
