@@ -12,9 +12,10 @@ import {
   ɵAngularFireSchedulers,
   ɵkeepUnstableUntilFirstFactory
 } from '@angular/fire';
-import { User, auth } from 'firebase/app';
+import firebase from 'firebase/app';
+import { isPlatformServer } from '@angular/common';
 
-export interface AngularFireAuth extends ɵPromiseProxy<auth.Auth> {}
+export interface AngularFireAuth extends ɵPromiseProxy<firebase.auth.Auth> {}
 
 @Injectable({
   providedIn: 'any'
@@ -24,7 +25,7 @@ export class AngularFireAuth {
   /**
    * Observable of authentication state; as of Firebase 4.0 this is only triggered via sign-in/out
    */
-  public readonly authState: Observable<User|null>;
+  public readonly authState: Observable<firebase.User|null>;
 
   /**
    * Observable of the currently signed-in user's JWT token used to identify the user to a Firebase service (or null).
@@ -34,14 +35,14 @@ export class AngularFireAuth {
   /**
    * Observable of the currently signed-in user (or null).
    */
-  public readonly user: Observable<User|null>;
+  public readonly user: Observable<firebase.User|null>;
 
   /**
    * Observable of the currently signed-in user's IdTokenResult object which contains the ID token JWT string and other
    * helper properties for getting different data associated with the token as well as all the decoded payload claims
    * (or null).
    */
-  public readonly idTokenResult: Observable<auth.IdTokenResult|null>;
+  public readonly idTokenResult: Observable<firebase.auth.IdTokenResult|null>;
 
   constructor(
     @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
@@ -61,30 +62,40 @@ export class AngularFireAuth {
       shareReplay({ bufferSize: 1, refCount: false }),
     );
 
-    // HACK, as we're exporting auth.Auth, rather than auth, developers importing firebase.auth
-    //       (e.g, `import { auth } from 'firebase/app'`) are getting an undefined auth object unexpectedly
-    //       as we're completely lazy. Let's eagerly load the Auth SDK here.
-    //       There could potentially be race conditions still... but this greatly decreases the odds while
-    //       we reevaluate the API.
-    const _ = auth.pipe(first()).subscribe();
+    if (isPlatformServer(platformId)) {
 
-    this.authState = auth.pipe(
-      switchMap(auth => zone.runOutsideAngular(() => new Observable<User|null>(auth.onAuthStateChanged.bind(auth)))),
-      keepUnstableUntilFirst
-    );
+      this.authState = this.user = this.idToken = this.idTokenResult = of(null);
 
-    this.user = auth.pipe(
-      switchMap(auth => zone.runOutsideAngular(() => new Observable<User|null>(auth.onIdTokenChanged.bind(auth)))),
-      keepUnstableUntilFirst
-    );
+    } else {
 
-    this.idToken = this.user.pipe(
-      switchMap(user => user ? from(user.getIdToken()) : of(null))
-    );
+      // HACK, as we're exporting auth.Auth, rather than auth, developers importing firebase.auth
+      //       (e.g, `import { auth } from 'firebase/app'`) are getting an undefined auth object unexpectedly
+      //       as we're completely lazy. Let's eagerly load the Auth SDK here.
+      //       There could potentially be race conditions still... but this greatly decreases the odds while
+      //       we reevaluate the API.
+      const _ = auth.pipe(first()).subscribe();
 
-    this.idTokenResult = this.user.pipe(
-      switchMap(user => user ? from(user.getIdTokenResult()) : of(null))
-    );
+      this.authState = auth.pipe(
+        switchMap(auth => auth.getRedirectResult().then(() => auth, () => auth)),
+        switchMap(auth => zone.runOutsideAngular(() => new Observable<firebase.User|null>(auth.onAuthStateChanged.bind(auth)))),
+        keepUnstableUntilFirst
+      );
+
+      this.user = auth.pipe(
+        switchMap(auth => auth.getRedirectResult().then(() => auth, () => auth)),
+        switchMap(auth => zone.runOutsideAngular(() => new Observable<firebase.User|null>(auth.onIdTokenChanged.bind(auth)))),
+        keepUnstableUntilFirst
+      );
+
+      this.idToken = this.user.pipe(
+        switchMap(user => user ? from(user.getIdToken()) : of(null))
+      );
+
+      this.idTokenResult = this.user.pipe(
+        switchMap(user => user ? from(user.getIdTokenResult()) : of(null))
+      );
+
+    }
 
     return ɵlazySDKProxy(this, auth, zone);
 
