@@ -1,4 +1,4 @@
-import { database } from 'firebase/app';
+import firebase from 'firebase/app';
 import { AngularFireModule, FirebaseApp } from '@angular/fire';
 import { AngularFireDatabase, AngularFireDatabaseModule, listChanges, URL } from '../public_api';
 import { TestBed } from '@angular/core/testing';
@@ -10,7 +10,7 @@ import { rando } from '../../firestore/utils.spec';
 describe('listChanges', () => {
   let app: FirebaseApp;
   let db: AngularFireDatabase;
-  let ref: (path: string) => database.Reference;
+  let ref: (path: string) => firebase.database.Reference;
   let batch = {};
   const items = [{ name: 'zero' }, { name: 'one' }, { name: 'two' }].map((item, i) => ({ key: i.toString(), ...item }));
   Object.keys(items).forEach((key, i) => {
