@@ -11,13 +11,13 @@ import {
   ɵlazySDKProxy,
   ɵPromiseProxy
 } from '@angular/fire';
-import { functions } from 'firebase/app';
+import firebase from 'firebase/app';
 
 export const ORIGIN = new InjectionToken<string>('angularfire2.functions.origin');
 export const REGION = new InjectionToken<string>('angularfire2.functions.region');
 
 // override httpsCallable for compatibility with 5.x
-export interface AngularFireFunctions extends Omit<ɵPromiseProxy<functions.Functions>, 'httpsCallable'> {
+export interface AngularFireFunctions extends Omit<ɵPromiseProxy<firebase.functions.Functions>, 'httpsCallable'> {
 }
 
 @Injectable({
@@ -51,7 +51,7 @@ export class AngularFireFunctions {
 
     this.httpsCallable = <T = any, R = any>(name: string) =>
       (data: T) => from(functions).pipe(
-        observeOn(schedulers.outsideAngular),
+        observeOn(schedulers.insideAngular),
         switchMap(functions => functions.httpsCallable(name)(data)),
         map(r => r.data as R)
       );
