@@ -26,17 +26,18 @@ import {
   ɵlazySDKProxy,
   ɵPromiseProxy
 } from '@angular/fire';
-import { remoteConfig } from 'firebase/app';
 import { isPlatformBrowser } from '@angular/common';
+import firebase from 'firebase/app';
+import { Settings } from './interfaces';
 
 export interface ConfigTemplate {
   [key: string]: string | number | boolean;
 }
 
-export const SETTINGS = new InjectionToken<remoteConfig.Settings>('angularfire2.remoteConfig.settings');
+export const SETTINGS = new InjectionToken<Settings>('angularfire2.remoteConfig.settings');
 export const DEFAULTS = new InjectionToken<ConfigTemplate>('angularfire2.remoteConfig.defaultConfig');
 
-export interface AngularFireRemoteConfig extends ɵPromiseProxy<remoteConfig.RemoteConfig> {
+export interface AngularFireRemoteConfig extends ɵPromiseProxy<firebase.remoteConfig.RemoteConfig> {
 }
 
 const AS_TO_FN = { strings: 'asString', numbers: 'asNumber', booleans: 'asBoolean' };
@@ -54,7 +55,7 @@ const proxyAll = (observable: Observable<Parameter[]>, as: 'numbers' | 'booleans
 ) as any;
 
 // TODO export as implements Partial<...> so minor doesn't break us
-export class Value implements remoteConfig.Value {
+export class Value implements firebase.remoteConfig.Value {
   asBoolean() {
     return ['1', 'true', 't', 'y', 'yes', 'on'].indexOf(this._value.toLowerCase()) > -1;
   }
@@ -72,13 +73,13 @@ export class Value implements remoteConfig.Value {
   }
 
   // tslint:disable-next-line:variable-name
-  constructor(public _source: remoteConfig.ValueSource, public _value: string) {
+  constructor(public _source: firebase.remoteConfig.ValueSource, public _value: string) {
   }
 }
 
 // SEMVER use ConstructorParameters when we can support Typescript 3.6
 export class Parameter extends Value {
-  constructor(public key: string, public fetchTimeMillis: number, source: remoteConfig.ValueSource, value: string) {
+  constructor(public key: string, public fetchTimeMillis: number, source: firebase.remoteConfig.ValueSource, value: string) {
     super(source, value);
   }
 }
@@ -99,8 +100,8 @@ export const filterFresh = (howRecentInMillis: number) => filterTest(p => p.fetc
 // on the Parameter. Also if it doesn't come from the server it won't emit again in .changes, due to the distinctUntilChanged,
 // which we can simplify to === rather than deep comparison
 const scanToParametersArray = (
-  remoteConfig: Observable<remoteConfig.RemoteConfig | undefined>
-): OperatorFunction<{ [key: string]: remoteConfig.Value }, Parameter[]> => pipe(
+  remoteConfig: Observable<firebase.remoteConfig.RemoteConfig | undefined>
+): OperatorFunction<{ [key: string]: firebase.remoteConfig.Value }, Parameter[]> => pipe(
   withLatestFrom(remoteConfig),
   scan((existing, [all, rc]) => {
     // SEMVER use "new Set" to unique once we're only targeting es6
@@ -131,7 +132,7 @@ export class AngularFireRemoteConfig {
   constructor(
     @Inject(FIREBASE_OPTIONS) options: FirebaseOptions,
     @Optional() @Inject(FIREBASE_APP_NAME) nameOrConfig: string | FirebaseAppConfig | null | undefined,
-    @Optional() @Inject(SETTINGS) settings: remoteConfig.Settings | null,
+    @Optional() @Inject(SETTINGS) settings: Settings | null,
     @Optional() @Inject(DEFAULTS) defaultConfig: ConfigTemplate | null,
     private zone: NgZone,
     // tslint:disable-next-line:ban-types
@@ -143,6 +144,8 @@ export class AngularFireRemoteConfig {
     const remoteConfig$ = of(undefined).pipe(
       observeOn(schedulers.outsideAngular),
       switchMap(() => isPlatformBrowser(platformId) ? import('firebase/remote-config') : EMPTY),
+      switchMap(() => import('@firebase/remote-config')),
+      tap(rc => rc.registerRemoteConfig && rc.registerRemoteConfig(firebase as any)),
       map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
       map(app => app.remoteConfig()),
       tap(rc => {
@@ -159,16 +162,16 @@ export class AngularFireRemoteConfig {
     );
 
     const loadedRemoteConfig$ = remoteConfig$.pipe(
-      filter<remoteConfig.RemoteConfig>(rc => !!rc)
+      filter<firebase.remoteConfig.RemoteConfig>(rc => !!rc)
     );
 
-    const default$: Observable<{ [key: string]: remoteConfig.Value }> = of(Object.keys(defaultConfig || {}).reduce(
+    const default$: Observable<{ [key: string]: firebase.remoteConfig.Value }> = of(Object.keys(defaultConfig || {}).reduce(
       (c, k) => ({ ...c, [k]: new Value('default', defaultConfig[k].toString()) }), {}
     ));
 
     // we should filter out the defaults we provided to RC, since we have our own implementation
     // that gives us a -1 for fetchTimeMillis (so filterFresh can filter them out)
-    const filterOutDefaults = map<{ [key: string]: remoteConfig.Value }, { [key: string]: remoteConfig.Value }>(all =>
+    const filterOutDefaults = map<{ [key: string]: firebase.remoteConfig.Value }, { [key: string]: firebase.remoteConfig.Value }>(all =>
       Object.keys(all)
         .filter(key => all[key].getSource() !== 'default')
         .reduce((acc, key) => ({ ...acc, [key]: all[key] }), {})
